requirements.txt
Flask==2.3.2
twilio==8.4.0
gspread==5.11.0
oauth2client==4.1.3
python-dotenv==1.0.0
regex==2023.12.25
spacy==3.5.3
en-core-web-sm==3.5.0
numpy==1.24.4
